(fix) Circular dependency
