+++ src/frontend/typer/dokumentasjon.ts
@@ -48,3 +48,26 @@ export enum EFiltyper {
     JPG = 'image/jpg',
     JPEG = 'image/jpeg',
 }
+
+export const dokumentasjonsbehovTilSpråkId = (dokumentasjonsbehov: Dokumentasjonsbehov): string => {
+    switch (dokumentasjonsbehov) {
+        case Dokumentasjonsbehov.ADOPSJON_DATO:
+            return 'dokumentasjon.adopsjon.vedleggtittel';
+        case Dokumentasjonsbehov.ANNEN_DOKUMENTASJON:
+            return 'dokumentasjon.annendokumentasjon.vedleggtittel';
+        case Dokumentasjonsbehov.AVTALE_DELT_BOSTED:
+            return 'dokumentasjon.deltbosted.vedleggtittel';
+        case Dokumentasjonsbehov.BEKREFTELSE_FRA_BARNEVERN:
+            return 'dokumentasjon.bekreftelsebarnevernet.vedleggtittel';
+        case Dokumentasjonsbehov.BOR_FAST_MED_SØKER:
+            return 'dokumentasjon.bekreftelseborsammen.vedleggtittel';
+        case Dokumentasjonsbehov.EØS_SKJEMA:
+            return 'dokumentasjon.tilleggsskjema.vedleggtittel';
+        case Dokumentasjonsbehov.MEKLINGSATTEST:
+            return 'dokumentasjon.meklingsattest.vedleggtittel';
+        case Dokumentasjonsbehov.SEPARERT_SKILT_ENKE:
+            return 'dokumentasjon.separasjonskilsmissedødsfall.vedleggtittel';
+        case Dokumentasjonsbehov.VEDTAK_OPPHOLDSTILLATELSE:
+            return 'dokumentasjon.oppholdstillatelse.vedleggtittel';
+    }
+};
