+++ src/frontend/utils/språk.ts
@@ -10,7 +10,6 @@ import * as engelsk from '../assets/lang/en.json';
 import * as bokmål from '../assets/lang/nb.json';
 import * as nynorsk from '../assets/lang/nn.json';
 import { innebygdeFormatterere } from '../components/Felleskomponenter/SpråkTekst/SpråkTekst';
-import { Dokumentasjonsbehov } from '../typer/dokumentasjon';
 import { AlternativtSvarForInput, ESivilstand, IBarn } from '../typer/person';
 import { Årsak } from '../typer/søknad';
 
@@ -35,29 +34,6 @@ export const toÅrsakSpråkId = (årsak: Årsak): string => {
 
 export const landkodeTilSpråk = (landkode: Alpha3Code | '', locale: string): string => {
     return landkode ? getName(alpha3ToAlpha2(landkode), locale) : AlternativtSvarForInput.UKJENT;
-};
-
-export const dokumentasjonsbehovTilSpråkId = (dokumentasjonsbehov: Dokumentasjonsbehov): string => {
-    switch (dokumentasjonsbehov) {
-        case Dokumentasjonsbehov.ADOPSJON_DATO:
-            return 'dokumentasjon.adopsjon.vedleggtittel';
-        case Dokumentasjonsbehov.ANNEN_DOKUMENTASJON:
-            return 'dokumentasjon.annendokumentasjon.vedleggtittel';
-        case Dokumentasjonsbehov.AVTALE_DELT_BOSTED:
-            return 'dokumentasjon.deltbosted.vedleggtittel';
-        case Dokumentasjonsbehov.BEKREFTELSE_FRA_BARNEVERN:
-            return 'dokumentasjon.bekreftelsebarnevernet.vedleggtittel';
-        case Dokumentasjonsbehov.BOR_FAST_MED_SØKER:
-            return 'dokumentasjon.bekreftelseborsammen.vedleggtittel';
-        case Dokumentasjonsbehov.EØS_SKJEMA:
-            return 'dokumentasjon.tilleggsskjema.vedleggtittel';
-        case Dokumentasjonsbehov.MEKLINGSATTEST:
-            return 'dokumentasjon.meklingsattest.vedleggtittel';
-        case Dokumentasjonsbehov.SEPARERT_SKILT_ENKE:
-            return 'dokumentasjon.separasjonskilsmissedødsfall.vedleggtittel';
-        case Dokumentasjonsbehov.VEDTAK_OPPHOLDSTILLATELSE:
-            return 'dokumentasjon.oppholdstillatelse.vedleggtittel';
-    }
 };
 
 const stripSpråkfil = (språkfilInnhold: Record<string, string>): Record<string, string> => {
